FunctionBench/cpu-memory/image_processing/requirements.txt
boto3
Pillow
scalene